# fix: jwt predicates with regexp cached on create.