fix: jwt predicates with regexp cached on create https://github.com/zalando/skipper/issues/3609.

Routing table updates are frequently in skipper. We assume every 3s a rebuild of the routing tree. Therefore a leak in objects like predicate creation can visibly leak memory. As written in https://github.com/zalando/skipper/issues/3609 and mentioned in https://github.com/zalando/skipper/pull/3600/files\#r2313466897 we appended unbounded and never cleand up predicates in the registry Now we use a map to also cache creation of the predicate to not grow unbounded and this also speedsup predicate creation and reduces number of objects, so allocations and memory if the same eskip.Predicate string would be used often.


Implementation of course do not change the hot path:
```
% benchstat old.txt new.txt                                                                                                                                  fix/predicate-cache-clean
goos: linux
goarch: amd64
pkg: github.com/zalando/skipper/predicates/auth
cpu: AMD Ryzen 7 PRO 4750U with Radeon Graphics
                        │   old.txt    │               new.txt               │
                        │    sec/op    │    sec/op     vs base               │
JWTPayloadAnyKVRegexp     88.05n ± 25%   87.00n ± 27%       ~ (p=0.726 n=10)
JWTPayloadAnyKVRegexp-2   66.28n ±  0%   66.61n ±  0%       ~ (p=0.093 n=10)
JWTPayloadAnyKVRegexp-4   66.52n ±  1%   66.41n ±  1%       ~ (p=0.325 n=10)
JWTPayloadAnyKVRegexp-8   66.56n ±  1%   66.71n ±  1%       ~ (p=0.956 n=10)
geomean                   71.30n         71.18n        -0.16%

                        │   old.txt    │               new.txt               │
                        │     B/op     │    B/op     vs base                 │
JWTPayloadAnyKVRegexp     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
JWTPayloadAnyKVRegexp-2   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
JWTPayloadAnyKVRegexp-4   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
JWTPayloadAnyKVRegexp-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                              ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                        │   old.txt    │               new.txt               │
                        │  allocs/op   │ allocs/op   vs base                 │
JWTPayloadAnyKVRegexp     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
JWTPayloadAnyKVRegexp-2   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
JWTPayloadAnyKVRegexp-4   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
JWTPayloadAnyKVRegexp-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                              ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```